+++ src/components/Scroll.js
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import ExpandLessIcon from "@material-ui/icons/ExpandLess";
-import IconButton from "@material-ui/core/IconButton";
 import Fab from "@material-ui/core/Fab";
 
 const useStyles = makeStyles((theme) => ({
