+++ src/components/SimpleTabs.js
@@ -1,8 +1,5 @@
 import React, { useState, useContext, useRef, useEffect } from "react";
 import { makeStyles, useTheme } from "@material-ui/core/styles";
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
-import Paper from "@material-ui/core/Paper";
 import Tabs from "@material-ui/core/Tabs";
 import Tab from "@material-ui/core/Tab";
 import Tooltip from "@material-ui/core/Tooltip";
@@ -10,15 +7,10 @@ import MapIcon from "@material-ui/icons/Map";
 import TableChart from "@material-ui/icons/TableChart";
 import TabPanel from "./TabPanel";
 import Map from "../Map";
-import Table from "../Table";
 import DevExTable from "../DevExTable";
 import LocationContext from "../LocationContext";
-import { Button } from "@material-ui/core";
-import AccountCircleIcon from "@material-ui/icons/AccountCircle";
-import IconButton from "@material-ui/core/IconButton";
 import StationList from "../pages/overview/StationList";
 import useMediaQuery from "@material-ui/core/useMediaQuery";
-import axios from "axios";
 import Scroll from "./Scroll";
 
 const useStyles = makeStyles((theme) => ({
@@ -32,11 +24,6 @@ const useStyles = makeStyles((theme) => ({
     right: theme.spacing(2),
   },
 }));
-
-const logout = () => {
-  let sessionUrl = "https://watsonc.admin.gc2.io/api/v2/session/stop";
-  return axios.get(sessionUrl);
-};
 
 const useElementVisible = (options) => {
   const containerRef = useRef(null);
@@ -59,9 +46,8 @@ const useElementVisible = (options) => {
   return [containerRef, isTabVisible];
 };
 
-export default function SimpleTabs(props) {
+export default function SimpleTabs() {
   const classes = useStyles();
-  const [value, setValue] = useState(0);
   const locationContext = useContext(LocationContext);
   const theme = useTheme();
   const matches = useMediaQuery(theme.breakpoints.down("sm"));
@@ -71,17 +57,9 @@ export default function SimpleTabs(props) {
     threshold: 1.0,
   });
 
-  const handleChange = (event, newValue) => {
-    //setValue(newValue);
+  const handleChange = (_, newValue) => {
     locationContext.setTabValue(newValue);
   };
-
-  const handleLogout = () => {
-    props.setToken(null);
-  };
-
-  const goToMap = () => setValue(1);
-  const goToTable = () => setValue(0);
 
   const KortIcon = (
     <Tooltip title='Kort'>
@@ -97,16 +75,6 @@ export default function SimpleTabs(props) {
 
   return (
     <div>
-      {/* <AppBar position='static' style={{ backgroundColor: "lightseagreen" }}>
-        
-      </AppBar> */}
-      {/* <AppBar position='static' style={{ backgroundColor: "lightseagreen" }}> */}
-      {/* <Toolbar style={{ flexGrow: 1, flexDirection: "row-reverse" }}>
-          <Button color='inherit' onClick={handleLogout}>
-            Logout
-          </Button>
-        </Toolbar> */}
-      {/* <AppBar position='sticky' style={{ backgroundColor: "white" }}> */}
       <Tabs
         value={locationContext.tabValue}
         onChange={handleChange}
@@ -118,13 +86,11 @@ export default function SimpleTabs(props) {
         <Tab icon={TableIcon} />
         <Tab icon={KortIcon} />
       </Tabs>
-      {/* </AppBar> */}
       <TabPanel value={locationContext.tabValue} index={0}>
-        {/* <DevExTable {...props} /> */}
-        {matches ? <StationList /> : <DevExTable {...props} />}
+        {matches ? <StationList /> : <DevExTable />}
       </TabPanel>
       <TabPanel value={locationContext.tabValue} index={1}>
-        <Map data={props.sensors} />
+        <Map />
       </TabPanel>
       {matches && !isTabVisible && (
         <Scroll scrollBelow={100} className={classes.fab} />
