+++ src/context/AuthContext.js
@@ -1,19 +1,19 @@
-import React from "react";
-import axios from "axios";
+// import React from "react";
+// import axios from "axios";
 
-const login = ({ email, password }) => {
-  let sessionUrl = "https://watsonc.admin.gc2.io/api/v2/session/start";
-  let _loginData = {
-    user: email,
-    password: password,
-    schema: null,
-  };
-  axios.post(sessionUrl, _loginData);
-};
+// const login = ({ email, password }) => {
+//   let sessionUrl = "https://watsonc.admin.gc2.io/api/v2/session/start";
+//   let _loginData = {
+//     user: email,
+//     password: password,
+//     schema: null,
+//   };
+//   axios.post(sessionUrl, _loginData);
+// };
 
-const logout = () => {
-  let sessionUrl = "https://watsonc.admin.gc2.io/api/v2/session/stop";
-  return axios.get(sessionUrl);
-};
+// const logout = () => {
+//   let sessionUrl = "https://watsonc.admin.gc2.io/api/v2/session/stop";
+//   return axios.get(sessionUrl);
+// };
 
-const getToken;
+// const getToken;
