+++ src/pages/Location/MinimalSelect.js
@@ -1,15 +1,12 @@
-import React, { useState, useContext } from "react";
+import React, { useState } from "react";
 import { useHistory } from "react-router-dom";
 import makeStyles from "@material-ui/core/styles/makeStyles";
-//import { useMinimalSelectStyles } from '@mui-treasury/styles/select/minimal';
 import minimalSelectStyles from "./minimalSelect.styles";
 import Select from "@material-ui/core/Select";
 import MenuItem from "@material-ui/core/MenuItem";
 import FormControl from "@material-ui/core/FormControl";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
-import LocationContext from "../../LocationContext";
 
-// Original design here: https://github.com/siriwatknp/mui-treasury/issues/540
 const useMinimalSelectStyles = makeStyles(minimalSelectStyles, {
   name: "MinimalSelect",
 });
@@ -19,22 +16,13 @@ const MinimalSelect = ({
   stationList,
   selectedStation,
   setSelectedItem,
-  setCurrStation,
-  currentStation,
 }) => {
   const [stationId, setStationId] = useState(selectedStation + "");
   const [isOpen, setIsOpen] = useState(true);
-  const context = useContext(LocationContext);
   const history = useHistory();
 
   const handleChange = (event) => {
-    // setStationId(event.target.value);
-    // console.log("stationId : ", event.target.value);
-    // console.log("stationList :", stationList);
-    // const station = currentStation(event.target.value, stationList);
-    // console.log("currentStation =>", station);
     setSelectedItem(event.target.value);
-    // setCurrStation(station);
     history.replace(`/location/${locid}/${event.target.value}`);
 
     setIsOpen(false);
@@ -87,15 +75,13 @@ const MinimalSelect = ({
         onOpen={handleOpen}
         onClose={handleClose}
       >
-        {/* <MenuItem value={0}>Principle</MenuItem>
-        <MenuItem value={1}>Sketch</MenuItem>
-        <MenuItem value={2}>Photoshop</MenuItem>
-        <MenuItem value={3}>Framer</MenuItem> */}
-        {stationList.map((station, index) => (
-          <MenuItem key={index} value={station.ts_id}>
-            {station.ts_name}
-          </MenuItem>
-        ))}
+        {stationList
+          .filter((t) => t.ts_name !== null)
+          .map((station) => (
+            <MenuItem key={station.ts_id} value={station.ts_id}>
+              {station.ts_name}
+            </MenuItem>
+          ))}
       </Select>
     </FormControl>
   );
