--- src/pages/Stamdata/table-data.js
@@ -1,2254 +1,0 @@
-export default {
-  success: "200",
-  result: [
-    {
-      stationid: 171,
-      stationname: "6.975_3",
-      parameter: "Vandstand",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "!",
-      locid: 333,
-    },
-    {
-      stationid: 121,
-      stationname: "Brinkhus1_1",
-      parameter: "Vandstand",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "!",
-      locid: 316,
-    },
-    {
-      stationid: 124,
-      stationname: "Brinkhus3_1",
-      parameter: "Vandstand",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "!",
-      locid: 318,
-    },
-    {
-      stationid: 173,
-      stationname: "Brinkhus_B1",
-      parameter: "Vandstand",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "!",
-      locid: 353,
-    },
-    {
-      stationid: 1033,
-      stationname: "07092-002 Klovtoft",
-      parameter: "Vandstand",
-      loc_owner: "Seges",
-      alarm: "!",
-      locid: 1306,
-    },
-    {
-      stationid: 129,
-      stationname: "34.2907_1",
-      parameter: "Vandstand",
-      loc_owner: "Aalborg Vand A/S",
-      alarm: "!",
-      locid: 321,
-    },
-    {
-      stationid: 141,
-      stationname: "DMR_S3",
-      parameter: "Vandstand",
-      loc_owner: "DMR",
-      alarm: "!",
-      locid: 339,
-    },
-    {
-      stationid: 142,
-      stationname: "DMR_S4",
-      parameter: "Vandstand",
-      loc_owner: "DMR",
-      alarm: "!",
-      locid: 340,
-    },
-    {
-      stationid: 132,
-      stationname: "B310",
-      parameter: "Vandstand",
-      loc_owner: "DMR",
-      alarm: "!",
-      locid: 324,
-    },
-    {
-      stationid: 181,
-      stationname: "Gråsten 2: Østersøen ved Slotsbakken",
-      parameter: "Vandstand",
-      loc_owner: "Sønderborg Kommune",
-      alarm: "!",
-      locid: 361,
-    },
-    {
-      stationid: 496,
-      stationname: "Glattrup Bæk",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "!",
-      locid: 672,
-    },
-    {
-      stationid: 505,
-      stationname: "Trævel Å, Kærgårdsholm",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "!",
-      locid: 681,
-    },
-    {
-      stationid: 585,
-      stationname: "Fårup rigkær",
-      parameter: "Vandstand",
-      loc_owner: "Randers Kommune",
-      alarm: "!",
-      locid: 740,
-    },
-    {
-      stationid: 625,
-      stationname: "Bastholm Mølle Å - Vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Hjørring Kommune",
-      alarm: "!",
-      locid: 769,
-    },
-    {
-      stationid: 563,
-      stationname: "Egetved H",
-      parameter: "Vandstand",
-      loc_owner: "Egetved",
-      alarm: "!",
-      locid: 727,
-    },
-    {
-      stationid: 679,
-      stationname: "Morsoegade_PB106-A_1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "!",
-      locid: 800,
-    },
-    {
-      stationid: 794,
-      stationname: "Hornslet, Engvej Flow",
-      parameter: "Flow",
-      loc_owner: "Syddjurs Spildevand",
-      alarm: "!",
-      locid: 847,
-    },
-    {
-      stationid: 793,
-      stationname: "Hornslet, Engvej Hastighed",
-      parameter: "Hastighed",
-      loc_owner: "Syddjurs Spildevand",
-      alarm: "!",
-      locid: 847,
-    },
-    {
-      stationid: 791,
-      stationname: "Hornslet, Engvej Vanddybde",
-      parameter: "Vandstand",
-      loc_owner: "Syddjurs Spildevand",
-      alarm: "!",
-      locid: 847,
-    },
-    {
-      stationid: 1136,
-      stationname: "Logger 498",
-      parameter: "Vandstand",
-      loc_owner: "PC",
-      alarm: "!",
-      locid: 1343,
-    },
-    {
-      stationid: 1130,
-      stationname: "Logger Flow",
-      parameter: "Flow",
-      loc_owner: "PC",
-      alarm: "!",
-      locid: 1343,
-    },
-    {
-      stationid: 798,
-      stationname: "PC4 (498)",
-      parameter: "Vandstand",
-      loc_owner: "PC Brøndboring",
-      alarm: "!",
-      locid: 849,
-    },
-    {
-      stationid: 805,
-      stationname: "Højgaard Fiskeri, indløb H",
-      parameter: "Vandstand",
-      loc_owner: "Højgård Fiskeri",
-      alarm: "!",
-      locid: 852,
-    },
-    {
-      stationid: 806,
-      stationname: "Højgaard Fiskeri, udløb H",
-      parameter: "Vandstand",
-      loc_owner: "Højgård Fiskeri",
-      alarm: "!",
-      locid: 853,
-    },
-    {
-      stationid: 1,
-      stationname: "Brabrand1_1",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 233,
-    },
-    {
-      stationid: 2,
-      stationname: "Brabrand1_2",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 233,
-    },
-    {
-      stationid: 3,
-      stationname: "Brabrand2_1",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 234,
-    },
-    {
-      stationid: 4,
-      stationname: "Brabrand2_2",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 234,
-    },
-    {
-      stationid: 5,
-      stationname: "Brabrand3_1",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 235,
-    },
-    {
-      stationid: 6,
-      stationname: "Brabrand3_2",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 235,
-    },
-    {
-      stationid: 7,
-      stationname: "Brabrand4_1",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 236,
-    },
-    {
-      stationid: 8,
-      stationname: "Brabrand4_2",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 236,
-    },
-    {
-      stationid: 172,
-      stationname: "6.557_1",
-      parameter: "Vandstand",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "OK",
-      locid: 329,
-    },
-    {
-      stationid: 153,
-      stationname: "6.966_2",
-      parameter: "Vandstand",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "OK",
-      locid: 330,
-    },
-    {
-      stationid: 174,
-      stationname: "6.972_2",
-      parameter: "Vandstand",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "OK",
-      locid: 354,
-    },
-    {
-      stationid: 156,
-      stationname: "6.975_1",
-      parameter: "Vandstand",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "OK",
-      locid: 333,
-    },
-    {
-      stationid: 122,
-      stationname: "Brinkhus1_2",
-      parameter: "Vandstand",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "OK",
-      locid: 316,
-    },
-    {
-      stationid: 123,
-      stationname: "Brinkhus2_1",
-      parameter: "Vandstand",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "OK",
-      locid: 317,
-    },
-    {
-      stationid: 161,
-      stationname: "BrinkhusQ1_Q",
-      parameter: "Vandføring",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "OK",
-      locid: 347,
-    },
-    {
-      stationid: 162,
-      stationname: "BrinkhusQ2_Q",
-      parameter: "Vandføring",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "OK",
-      locid: 348,
-    },
-    {
-      stationid: 154,
-      stationname: "Nedbør Brinkhus",
-      parameter: "Nedbør",
-      loc_owner: "Frederikshavn Vand A/S",
-      alarm: "OK",
-      locid: 332,
-    },
-    {
-      stationid: 157,
-      stationname: "Brokær23_1",
-      parameter: "Vandstand",
-      loc_owner: "Thisted Kommune",
-      alarm: "OK",
-      locid: 335,
-    },
-    {
-      stationid: 158,
-      stationname: "Brokær24_1",
-      parameter: "Vandstand",
-      loc_owner: "Thisted Kommune",
-      alarm: "OK",
-      locid: 336,
-    },
-    {
-      stationid: 18,
-      stationname: "Brokær5_1",
-      parameter: "Vandstand",
-      loc_owner: "Thisted Kommune",
-      alarm: "OK",
-      locid: 251,
-    },
-    {
-      stationid: 30,
-      stationname: "Glombak4_1",
-      parameter: "Vandstand",
-      loc_owner: "Thisted Kommune",
-      alarm: "OK",
-      locid: 289,
-    },
-    {
-      stationid: 31,
-      stationname: "Glombak4_2",
-      parameter: "Vandstand",
-      loc_owner: "Thisted Kommune",
-      alarm: "OK",
-      locid: 289,
-    },
-    {
-      stationid: 60,
-      stationname: "Nørhå5_1",
-      parameter: "Vandstand",
-      loc_owner: "Thisted Kommune",
-      alarm: "OK",
-      locid: 275,
-    },
-    {
-      stationid: 62,
-      stationname: "Nørhå5_2",
-      parameter: "Vandstand",
-      loc_owner: "Thisted Kommune",
-      alarm: "OK",
-      locid: 275,
-    },
-    {
-      stationid: 175,
-      stationname: "SønderHå9_1",
-      parameter: "Vandstand",
-      loc_owner: "Thisted Kommune",
-      alarm: "OK",
-      locid: 355,
-    },
-    {
-      stationid: 601,
-      stationname: "Throughflow_1",
-      parameter: "Nedbør",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 751,
-    },
-    {
-      stationid: 602,
-      stationname: "Throughflow_2",
-      parameter: "Nedbør",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 752,
-    },
-    {
-      stationid: 603,
-      stationname: "Throughflow_3",
-      parameter: "Nedbør",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 753,
-    },
-    {
-      stationid: 728,
-      stationname: "Vårst_Fugtighed_1_1",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 751,
-    },
-    {
-      stationid: 729,
-      stationname: "Vårst_Fugtighed_1_2",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 751,
-    },
-    {
-      stationid: 730,
-      stationname: "Vårst_Fugtighed_1_3",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 751,
-    },
-    {
-      stationid: 731,
-      stationname: "Vårst_Fugtighed_1_4",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 751,
-    },
-    {
-      stationid: 732,
-      stationname: "Vårst_Fugtighed_1_5",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 751,
-    },
-    {
-      stationid: 733,
-      stationname: "Vårst_Fugtighed_1_6",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 751,
-    },
-    {
-      stationid: 734,
-      stationname: "Vårst_Fugtighed_2_1",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 752,
-    },
-    {
-      stationid: 735,
-      stationname: "Vårst_Fugtighed_2_2",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 752,
-    },
-    {
-      stationid: 736,
-      stationname: "Vårst_Fugtighed_2_3",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 752,
-    },
-    {
-      stationid: 737,
-      stationname: "Vårst_Fugtighed_2_4",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 752,
-    },
-    {
-      stationid: 738,
-      stationname: "Vårst_Fugtighed_2_5",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 752,
-    },
-    {
-      stationid: 739,
-      stationname: "Vårst_Fugtighed_2_6",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 752,
-    },
-    {
-      stationid: 740,
-      stationname: "Vårst_Fugtighed_3_1",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 753,
-    },
-    {
-      stationid: 741,
-      stationname: "Vårst_Fugtighed_3_2",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 753,
-    },
-    {
-      stationid: 742,
-      stationname: "Vårst_Fugtighed_3_3",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 753,
-    },
-    {
-      stationid: 743,
-      stationname: "Vårst_Fugtighed_3_4",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 753,
-    },
-    {
-      stationid: 744,
-      stationname: "Vårst_Fugtighed_3_5",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 753,
-    },
-    {
-      stationid: 745,
-      stationname: "Vårst_Fugtighed_3_6",
-      parameter: "Fugtighed",
-      loc_owner: "Aalborg Forsyning",
-      alarm: "OK",
-      locid: 753,
-    },
-    {
-      stationid: 727,
-      stationname: "Vårst_Nedbør",
-      parameter: "Nedbør",
-      loc_owner: "undefined",
-      alarm: "OK",
-      locid: 834,
-    },
-    {
-      stationid: 170,
-      stationname: "07014 Ringsted",
-      parameter: "Vandstand",
-      loc_owner: "Seges",
-      alarm: "OK",
-      locid: 352,
-    },
-    {
-      stationid: 177,
-      stationname: "07027 Bur",
-      parameter: "Vandstand",
-      loc_owner: "Seges",
-      alarm: "OK",
-      locid: 357,
-    },
-    {
-      stationid: 116,
-      stationname: "07028_4b Guldborg",
-      parameter: "Vandstand",
-      loc_owner: "Seges",
-      alarm: "OK",
-      locid: 314,
-    },
-    {
-      stationid: 1034,
-      stationname: "07092-002 Rebbøl",
-      parameter: "Vandstand",
-      loc_owner: "Seges",
-      alarm: "OK",
-      locid: 1307,
-    },
-    {
-      stationid: 169,
-      stationname: "07093 Hellevad",
-      parameter: "Vandstand",
-      loc_owner: "Seges",
-      alarm: "OK",
-      locid: 351,
-    },
-    {
-      stationid: 1032,
-      stationname: "07093 Rangstrup",
-      parameter: "Vandstand",
-      loc_owner: "Seges",
-      alarm: "OK",
-      locid: 1305,
-    },
-    {
-      stationid: 115,
-      stationname: "07096_3b Odder",
-      parameter: "Vandstand",
-      loc_owner: "Seges",
-      alarm: "OK",
-      locid: 313,
-    },
-    {
-      stationid: 130,
-      stationname: "34.1654_1",
-      parameter: "Vandstand",
-      loc_owner: "Aalborg Vand A/S",
-      alarm: "OK",
-      locid: 322,
-    },
-    {
-      stationid: 125,
-      stationname: "34.4408 _1",
-      parameter: "Vandstand",
-      loc_owner: "Aalborg Vand A/S",
-      alarm: "OK",
-      locid: 319,
-    },
-    {
-      stationid: 126,
-      stationname: "34.4408 _2",
-      parameter: "Vandstand",
-      loc_owner: "Aalborg Vand A/S",
-      alarm: "OK",
-      locid: 319,
-    },
-    {
-      stationid: 127,
-      stationname: "34.4409_1",
-      parameter: "Vandstand",
-      loc_owner: "Aalborg Vand A/S",
-      alarm: "OK",
-      locid: 320,
-    },
-    {
-      stationid: 128,
-      stationname: "34.4409_2",
-      parameter: "Vandstand",
-      loc_owner: "Aalborg Vand A/S",
-      alarm: "OK",
-      locid: 320,
-    },
-    {
-      stationid: 462,
-      stationname: "SofiendalEnge5_1",
-      parameter: "Vandstand",
-      loc_owner: "Aalborg Vand A/S",
-      alarm: "OK",
-      locid: 639,
-    },
-    {
-      stationid: 134,
-      stationname: "DMR_N1",
-      parameter: "Vandstand",
-      loc_owner: "DMR",
-      alarm: "OK",
-      locid: 342,
-    },
-    {
-      stationid: 139,
-      stationname: "DMR_N2",
-      parameter: "Vandstand",
-      loc_owner: "DMR",
-      alarm: "OK",
-      locid: 343,
-    },
-    {
-      stationid: 137,
-      stationname: "DMR_N4",
-      parameter: "Vandstand",
-      loc_owner: "DMR",
-      alarm: "OK",
-      locid: 345,
-    },
-    {
-      stationid: 138,
-      stationname: "DMR_N5",
-      parameter: "Vandstand",
-      loc_owner: "DMR",
-      alarm: "OK",
-      locid: 346,
-    },
-    {
-      stationid: 135,
-      stationname: "DMR_S1",
-      parameter: "Vandstand",
-      loc_owner: "DMR",
-      alarm: "OK",
-      locid: 337,
-    },
-    {
-      stationid: 140,
-      stationname: "DMR_S2",
-      parameter: "Vandstand",
-      loc_owner: "DMR",
-      alarm: "OK",
-      locid: 338,
-    },
-    {
-      stationid: 131,
-      stationname: "B202",
-      parameter: "Vandstand",
-      loc_owner: "DMR",
-      alarm: "OK",
-      locid: 323,
-    },
-    {
-      stationid: 133,
-      stationname: "B313",
-      parameter: "Vandstand",
-      loc_owner: "DMR",
-      alarm: "OK",
-      locid: 325,
-    },
-    {
-      stationid: 180,
-      stationname: "Gråsten 1: Slotssøen",
-      parameter: "Vandstand",
-      loc_owner: "Sønderborg Kommune",
-      alarm: "OK",
-      locid: 360,
-    },
-    {
-      stationid: 182,
-      stationname: "Gråsten 3: Østersøen opstrøms Kongevej",
-      parameter: "Vandstand",
-      loc_owner: "Sønderborg Kommune",
-      alarm: "OK",
-      locid: 362,
-    },
-    {
-      stationid: 183,
-      stationname: "Gråsten 4: Østersøen nedstrøms Kongevej",
-      parameter: "Vandstand",
-      loc_owner: "Sønderborg Kommune",
-      alarm: "OK",
-      locid: 363,
-    },
-    {
-      stationid: 184,
-      stationname: "Gråsten 5: Østersøen ved Fiskenæspumpen",
-      parameter: "Vandstand",
-      loc_owner: "Sønderborg Kommune",
-      alarm: "OK",
-      locid: 364,
-    },
-    {
-      stationid: 185,
-      stationname: "Gråsten 6: Havet ved Østersøvej",
-      parameter: "Vandstand",
-      loc_owner: "Sønderborg Kommune",
-      alarm: "OK",
-      locid: 365,
-    },
-    {
-      stationid: 543,
-      stationname: "Kær Vig",
-      parameter: "Vandstand",
-      loc_owner: "Sønderborg Kommune",
-      alarm: "OK",
-      locid: 719,
-    },
-    {
-      stationid: 578,
-      stationname: "Skovmose",
-      parameter: null,
-      loc_owner: "Sønderborg Kommune",
-      alarm: "OK",
-      locid: 736,
-    },
-    {
-      stationid: 178,
-      stationname: "Koholm Å, Estvadgårds Bro",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 358,
-    },
-    {
-      stationid: 189,
-      stationname: "75.2156_1 (B3)",
-      parameter: "Vandstand",
-      loc_owner: "GEUS",
-      alarm: "OK",
-      locid: 367,
-    },
-    {
-      stationid: 190,
-      stationname: "75.2160_1 (B7)",
-      parameter: "Vandstand",
-      loc_owner: "GEUS",
-      alarm: "OK",
-      locid: 368,
-    },
-    {
-      stationid: 210,
-      stationname: "Brande_B1",
-      parameter: "Vandstand",
-      loc_owner: "Envidan",
-      alarm: "OK",
-      locid: 635,
-    },
-    {
-      stationid: 208,
-      stationname: "Brande_B2",
-      parameter: "Vandstand",
-      loc_owner: "Envidan",
-      alarm: "OK",
-      locid: 637,
-    },
-    {
-      stationid: 209,
-      stationname: "Brande_B3",
-      parameter: "Vandstand",
-      loc_owner: "Envidan",
-      alarm: "OK",
-      locid: 636,
-    },
-    {
-      stationid: 211,
-      stationname: "Brande_B4",
-      parameter: "Vandstand",
-      loc_owner: "Envidan",
-      alarm: "OK",
-      locid: 634,
-    },
-    {
-      stationid: 207,
-      stationname: "Brande_nedbør",
-      parameter: "Nedbør",
-      loc_owner: "Envidan",
-      alarm: "OK",
-      locid: 638,
-    },
-    {
-      stationid: 199,
-      stationname: "GranAlle_1",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Forsyning",
-      alarm: "OK",
-      locid: 375,
-    },
-    {
-      stationid: 200,
-      stationname: "Lupinvej_1",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Forsyning",
-      alarm: "OK",
-      locid: 376,
-    },
-    {
-      stationid: 692,
-      stationname: "NordreHøjmarksvej",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Forsyning",
-      alarm: "OK",
-      locid: 811,
-    },
-    {
-      stationid: 694,
-      stationname: "Purhusvej",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Forsyning",
-      alarm: "OK",
-      locid: 813,
-    },
-    {
-      stationid: 693,
-      stationname: "Skærskovgårdvej",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Forsyning",
-      alarm: "OK",
-      locid: 812,
-    },
-    {
-      stationid: 691,
-      stationname: "Skærskovvej",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Forsyning",
-      alarm: "OK",
-      locid: 810,
-    },
-    {
-      stationid: 201,
-      stationname: "Sportsvej 1",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Forsyning",
-      alarm: "OK",
-      locid: 377,
-    },
-    {
-      stationid: 202,
-      stationname: "Tisdalen_1",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Forsyning",
-      alarm: "OK",
-      locid: 378,
-    },
-    {
-      stationid: 495,
-      stationname: "Bærs Bæk",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 671,
-    },
-    {
-      stationid: 497,
-      stationname: "Grynderup Sø, Afløb",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 673,
-    },
-    {
-      stationid: 498,
-      stationname: "Hinnerup Å, Gl. Åstedvej",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 674,
-    },
-    {
-      stationid: 499,
-      stationname: "Hinnerup Å, Østergård",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 675,
-    },
-    {
-      stationid: 577,
-      stationname: "Kåstrup Bæk",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 735,
-    },
-    {
-      stationid: 500,
-      stationname: "Langvad Å",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 676,
-    },
-    {
-      stationid: 576,
-      stationname: "Næstild Bæk",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 734,
-    },
-    {
-      stationid: 501,
-      stationname: "Rødding Å",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 677,
-    },
-    {
-      stationid: 179,
-      stationname: "Skive Å, Engvej",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 359,
-    },
-    {
-      stationid: 502,
-      stationname: "Skive Å, Lundbro",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 678,
-    },
-    {
-      stationid: 503,
-      stationname: "Sæby Bæk Bysted Bro",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 679,
-    },
-    {
-      stationid: 504,
-      stationname: "Trævel Å, Håsum Enge",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 680,
-    },
-    {
-      stationid: 587,
-      stationname: "Ansø_Rigkær1_1",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Kommune",
-      alarm: "OK",
-      locid: 742,
-    },
-    {
-      stationid: 588,
-      stationname: "Ansø_Rigkær1_2",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Kommune",
-      alarm: "OK",
-      locid: 742,
-    },
-    {
-      stationid: 589,
-      stationname: "Ansø_Rigkær2",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Kommune",
-      alarm: "OK",
-      locid: 743,
-    },
-    {
-      stationid: 590,
-      stationname: "Ansø_Rigkær3",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Kommune",
-      alarm: "OK",
-      locid: 744,
-    },
-    {
-      stationid: 586,
-      stationname: "Salten Å - Ansøvej",
-      parameter: "Vandstand",
-      loc_owner: "Silkeborg Kommune",
-      alarm: "OK",
-      locid: 741,
-    },
-    {
-      stationid: 506,
-      stationname: "38.628_1",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 682,
-    },
-    {
-      stationid: 510,
-      stationname: "35.869_1",
-      parameter: "Vandstand",
-      loc_owner: "VSK",
-      alarm: "OK",
-      locid: 686,
-    },
-    {
-      stationid: 541,
-      stationname: "Ferup_Sø",
-      parameter: "Vandstand",
-      loc_owner: "Trefor",
-      alarm: "OK",
-      locid: 717,
-    },
-    {
-      stationid: 540,
-      stationname: "Harteværket",
-      parameter: "Vandstand",
-      loc_owner: "Trefor",
-      alarm: "OK",
-      locid: 716,
-    },
-    {
-      stationid: 556,
-      stationname: "Felding Bæk Havnevej",
-      parameter: "Vandstand",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 732,
-    },
-    {
-      stationid: 584,
-      stationname: "Brande mose, Hyvildvej",
-      parameter: "Vandstand",
-      loc_owner: "Ikast/Brande Kommune",
-      alarm: "OK",
-      locid: 739,
-    },
-    {
-      stationid: 612,
-      stationname: "Blåsig Bæk, Frh.Havnsvej - Vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 762,
-    },
-    {
-      stationid: 613,
-      stationname: "Liver Å, Grimsted bro - Vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 763,
-    },
-    {
-      stationid: 1039,
-      stationname: "Liver Å, Grimsted bro - Iltindhold",
-      parameter: "Opløst ilt",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 763,
-    },
-    {
-      stationid: 1038,
-      stationname: "Liver Å, Grimsted bro - Temperatur",
-      parameter: "Temperatur",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 763,
-    },
-    {
-      stationid: 614,
-      stationname: "Liver Å, Vejby bro - Vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 764,
-    },
-    {
-      stationid: 624,
-      stationname: "Rakkeby Å, Ådalen - Vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 768,
-    },
-    {
-      stationid: 627,
-      stationname: "Tryn Å - Vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 771,
-    },
-    {
-      stationid: 1041,
-      stationname: "Uggerby Å, Ilbro - Iltindhold",
-      parameter: "Opløst ilt",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 767,
-    },
-    {
-      stationid: 1040,
-      stationname: "Uggerby Å, Ilbro - Temperatur",
-      parameter: "Temperatur",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 767,
-    },
-    {
-      stationid: 617,
-      stationname: "Uggerby Å, Ilbro - Vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 767,
-    },
-    {
-      stationid: 1043,
-      stationname: "Uggerby Å, Sindal - Iltindhold",
-      parameter: "Opløst ilt",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 765,
-    },
-    {
-      stationid: 1042,
-      stationname: "Uggerby Å, Sindal - Temperatur",
-      parameter: "Temperatur",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 765,
-    },
-    {
-      stationid: 615,
-      stationname: "Uggerby Å, Sindal - Vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 765,
-    },
-    {
-      stationid: 616,
-      stationname: "Uggerby Å, Sønderskov - Vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 766,
-    },
-    {
-      stationid: 628,
-      stationname: "Varbro Å - Vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 772,
-    },
-    {
-      stationid: 626,
-      stationname: "Vien Bæk - Vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Hjørring Kommune",
-      alarm: "OK",
-      locid: 770,
-    },
-    {
-      stationid: 559,
-      stationname: "Ølufvad indløb H",
-      parameter: "Vandstand",
-      loc_owner: "Ølufvad",
-      alarm: "OK",
-      locid: 722,
-    },
-    {
-      stationid: 557,
-      stationname: "Ølufvad udløb H",
-      parameter: "Vandstand",
-      loc_owner: "Ølufvad",
-      alarm: "OK",
-      locid: 725,
-    },
-    {
-      stationid: 564,
-      stationname: "Fårup H",
-      parameter: "Vandstand",
-      loc_owner: "Aquasearch",
-      alarm: "OK",
-      locid: 728,
-    },
-    {
-      stationid: 562,
-      stationname: "Hallesø H",
-      parameter: "Vandstand",
-      loc_owner: "Aquasearch",
-      alarm: "OK",
-      locid: 726,
-    },
-    {
-      stationid: 560,
-      stationname: "Hallesø Huse H",
-      parameter: "Vandstand",
-      loc_owner: "Aquasearch",
-      alarm: "OK",
-      locid: 723,
-    },
-    {
-      stationid: 558,
-      stationname: "Refsgaard H",
-      parameter: "Vandstand",
-      loc_owner: "Aquasearch",
-      alarm: "OK",
-      locid: 721,
-    },
-    {
-      stationid: 561,
-      stationname: "Brogård H",
-      parameter: "Vandstand",
-      loc_owner: "Frea",
-      alarm: "OK",
-      locid: 724,
-    },
-    {
-      stationid: 598,
-      stationname: "Kolding_Rigkær1_1",
-      parameter: "Vandstand",
-      loc_owner: "Kolding Kommune",
-      alarm: "OK",
-      locid: 749,
-    },
-    {
-      stationid: 597,
-      stationname: "Kolding_Rigkær2_1",
-      parameter: "Vandstand",
-      loc_owner: "Kolding Kommune",
-      alarm: "OK",
-      locid: 748,
-    },
-    {
-      stationid: 595,
-      stationname: "Kolding_Rigkær4_1",
-      parameter: "Vandstand",
-      loc_owner: "Kolding Kommune",
-      alarm: "OK",
-      locid: 747,
-    },
-    {
-      stationid: 596,
-      stationname: "Kolding_Rigkær4_2",
-      parameter: "Vandstand",
-      loc_owner: "Kolding Kommune",
-      alarm: "OK",
-      locid: 747,
-    },
-    {
-      stationid: 594,
-      stationname: "Kolding_Å",
-      parameter: "Vandstand",
-      loc_owner: "Kolding Kommune",
-      alarm: "OK",
-      locid: 746,
-    },
-    {
-      stationid: 698,
-      stationname: "DGU41.2113_2",
-      parameter: "Vandstand",
-      loc_owner: "Skørping Vandværk",
-      alarm: "OK",
-      locid: 817,
-    },
-    {
-      stationid: 699,
-      stationname: "DGU41.2114_1",
-      parameter: "Vandstand",
-      loc_owner: "Skørping Vandværk",
-      alarm: "OK",
-      locid: 818,
-    },
-    {
-      stationid: 604,
-      stationname: "Hjordkær_1",
-      parameter: "Vandstand",
-      loc_owner: "Aabenra Kommune",
-      alarm: "OK",
-      locid: 754,
-    },
-    {
-      stationid: 605,
-      stationname: "Hjordkær_2",
-      parameter: "Vandstand",
-      loc_owner: "Aabenra Kommune",
-      alarm: "OK",
-      locid: 755,
-    },
-    {
-      stationid: 606,
-      stationname: "Hjordkær_3",
-      parameter: "Vandstand",
-      loc_owner: "Aabenra Kommune",
-      alarm: "OK",
-      locid: 756,
-    },
-    {
-      stationid: 607,
-      stationname: "Hjordkær_4",
-      parameter: "Vandstand",
-      loc_owner: "Aabenra Kommune",
-      alarm: "OK",
-      locid: 757,
-    },
-    {
-      stationid: 608,
-      stationname: "Hjordkær_5",
-      parameter: "Vandstand",
-      loc_owner: "Aabenra Kommune",
-      alarm: "OK",
-      locid: 758,
-    },
-    {
-      stationid: 609,
-      stationname: "Hjordkær_6",
-      parameter: "Vandstand",
-      loc_owner: "Aabenra Kommune",
-      alarm: "OK",
-      locid: 759,
-    },
-    {
-      stationid: 610,
-      stationname: "Hjordkær_7",
-      parameter: "Vandstand",
-      loc_owner: "Aabenra Kommune",
-      alarm: "OK",
-      locid: 760,
-    },
-    {
-      stationid: 611,
-      stationname: "Hjordkær_8",
-      parameter: "Vandstand",
-      loc_owner: "Aabenra Kommune",
-      alarm: "OK",
-      locid: 761,
-    },
-    {
-      stationid: 803,
-      stationname: "Fremtidens_B2",
-      parameter: null,
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 850,
-    },
-    {
-      stationid: 804,
-      stationname: "Fremtidens_B3",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 851,
-    },
-    {
-      stationid: 684,
-      stationname: "Fremtidens_F083_1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 805,
-    },
-    {
-      stationid: 685,
-      stationname: "Fremtidens_F090_1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 806,
-    },
-    {
-      stationid: 686,
-      stationname: "Fremtidens_GB103_1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 807,
-    },
-    {
-      stationid: 751,
-      stationname: "Holstebro_B1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 782,
-    },
-    {
-      stationid: 754,
-      stationname: "Holstebro_B10.3",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 782,
-    },
-    {
-      stationid: 755,
-      stationname: "Holstebro_B13",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 782,
-    },
-    {
-      stationid: 752,
-      stationname: "Holstebro_B5",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 782,
-    },
-    {
-      stationid: 753,
-      stationname: "Holstebro_B9.3",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 782,
-    },
-    {
-      stationid: 680,
-      stationname: "Morsoegade_PB106-B_1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 801,
-    },
-    {
-      stationid: 1035,
-      stationname: "Morsoegade_PB11A",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 1308,
-    },
-    {
-      stationid: 1036,
-      stationname: "Morsoegade_PB12A",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 1309,
-    },
-    {
-      stationid: 1037,
-      stationname: "Morsoegade_PB13A",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 1310,
-    },
-    {
-      stationid: 1023,
-      stationname: "Vejle_MP1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 1296,
-    },
-    {
-      stationid: 1024,
-      stationname: "Vejle_MP2",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 1297,
-    },
-    {
-      stationid: 1025,
-      stationname: "Vejle_MP3",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 1298,
-    },
-    {
-      stationid: 672,
-      stationname: "Possø_grundvand",
-      parameter: "Vandstand",
-      loc_owner: "Sweco",
-      alarm: "OK",
-      locid: 795,
-    },
-    {
-      stationid: 700,
-      stationname: "Possø_nedbør",
-      parameter: "Nedbør",
-      loc_owner: "Sweco",
-      alarm: "OK",
-      locid: 819,
-    },
-    {
-      stationid: 671,
-      stationname: "Possø vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Sweco",
-      alarm: "OK",
-      locid: 794,
-    },
-    {
-      stationid: 634,
-      stationname: "Glenstrup Sø, Vest",
-      parameter: "Vandstand",
-      loc_owner: "Mariagerfjord Kommune",
-      alarm: "OK",
-      locid: 773,
-    },
-    {
-      stationid: 637,
-      stationname: "Glerup Bæk",
-      parameter: "Vandstand",
-      loc_owner: "Mariagerfjord Kommune",
-      alarm: "OK",
-      locid: 776,
-    },
-    {
-      stationid: 635,
-      stationname: "Kastbjerg Å",
-      parameter: "Vandstand",
-      loc_owner: "Mariagerfjord Kommune",
-      alarm: "OK",
-      locid: 774,
-    },
-    {
-      stationid: 636,
-      stationname: "Korup Å",
-      parameter: "Vandstand",
-      loc_owner: "Mariagerfjord Kommune",
-      alarm: "OK",
-      locid: 775,
-    },
-    {
-      stationid: 687,
-      stationname: "Kasted1_1",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 808,
-    },
-    {
-      stationid: 688,
-      stationname: "Kasted1_2",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 808,
-    },
-    {
-      stationid: 689,
-      stationname: "Kasted2_1",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 809,
-    },
-    {
-      stationid: 690,
-      stationname: "Kasted2_2",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 809,
-    },
-    {
-      stationid: 1141,
-      stationname: "Kasted3_1",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 1345,
-    },
-    {
-      stationid: 1142,
-      stationname: "Kasted3_2",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 1345,
-    },
-    {
-      stationid: 1143,
-      stationname: "Kasted4_1",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 1346,
-    },
-    {
-      stationid: 1144,
-      stationname: "Kasted4_2",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 1346,
-    },
-    {
-      stationid: 697,
-      stationname: "Lyngby Sø",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 816,
-    },
-    {
-      stationid: 673,
-      stationname: "Tingvad1_1",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 796,
-    },
-    {
-      stationid: 674,
-      stationname: "Tingvad1_2",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 796,
-    },
-    {
-      stationid: 675,
-      stationname: "Tingvad2_1",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 797,
-    },
-    {
-      stationid: 676,
-      stationname: "Tingvad2_2",
-      parameter: "Vandstand",
-      loc_owner: "Aarhus Vand",
-      alarm: "OK",
-      locid: 797,
-    },
-    {
-      stationid: 807,
-      stationname: "DGU 78.2403_1",
-      parameter: "Vandstand",
-      loc_owner: "Nymølle Stenindustri",
-      alarm: "OK",
-      locid: 854,
-    },
-    {
-      stationid: 668,
-      stationname: "Karup Å - Rigkær",
-      parameter: "Vandstand",
-      loc_owner: "COWI",
-      alarm: "OK",
-      locid: 793,
-    },
-    {
-      stationid: 600,
-      stationname: "Esbjerg Lufthavn B1-2",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 750,
-    },
-    {
-      stationid: 809,
-      stationname: "Esbjerg Strandpark  A1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 856,
-    },
-    {
-      stationid: 810,
-      stationname: "Esbjerg Strandpark  A2",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 857,
-    },
-    {
-      stationid: 1016,
-      stationname: "Esbjerg Strandpark B1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 1291,
-    },
-    {
-      stationid: 1017,
-      stationname: "Esbjerg Strandpark B2",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 1291,
-    },
-    {
-      stationid: 1018,
-      stationname: "Esbjerg Strandpark C1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 1292,
-    },
-    {
-      stationid: 1019,
-      stationname: "Esbjerg Strandpark C2",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 1292,
-    },
-    {
-      stationid: 1020,
-      stationname: " Langdalshøjen, Sejs",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 1293,
-    },
-    {
-      stationid: 649,
-      stationname: "Rambøll_Reserve1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 782,
-    },
-    {
-      stationid: 695,
-      stationname: "OdenseAa_Bellinge_Pb1",
-      parameter: "Vandstand",
-      loc_owner: "Energinet",
-      alarm: "OK",
-      locid: 814,
-    },
-    {
-      stationid: 696,
-      stationname: "OdenseAa_Bellinge_Pb2",
-      parameter: "Vandstand",
-      loc_owner: "Energinet",
-      alarm: "OK",
-      locid: 815,
-    },
-    {
-      stationid: 788,
-      stationname: "Brande Å",
-      parameter: "Vandstand",
-      loc_owner: "Ikast Brande Kommune",
-      alarm: "OK",
-      locid: 844,
-    },
-    {
-      stationid: 790,
-      stationname: "Karup Å",
-      parameter: "Vandstand",
-      loc_owner: "Ikast Brande Kommune",
-      alarm: "OK",
-      locid: 846,
-    },
-    {
-      stationid: 787,
-      stationname: "Mattrup Å",
-      parameter: "Vandstand",
-      loc_owner: "Ikast Brande Kommune",
-      alarm: "OK",
-      locid: 843,
-    },
-    {
-      stationid: 789,
-      stationname: "Storå",
-      parameter: "Vandstand",
-      loc_owner: "Ikast Brande Kommune",
-      alarm: "OK",
-      locid: 845,
-    },
-    {
-      stationid: 722,
-      stationname: "HOFOR Movable 1",
-      parameter: "Vandstand",
-      loc_owner: "undefined",
-      alarm: "OK",
-      locid: 832,
-    },
-    {
-      stationid: 748,
-      stationname: "HOFOR Movable 2",
-      parameter: "Vandstand",
-      loc_owner: "undefined",
-      alarm: "OK",
-      locid: 832,
-    },
-    {
-      stationid: 749,
-      stationname: "HOFOR Movable 3",
-      parameter: "Vandstand",
-      loc_owner: "undefined",
-      alarm: "OK",
-      locid: 832,
-    },
-    {
-      stationid: 750,
-      stationname: "HOFOR Movable 4",
-      parameter: "Vandstand",
-      loc_owner: "undefined",
-      alarm: "OK",
-      locid: 832,
-    },
-    {
-      stationid: 1014,
-      stationname: "Hinnerup Å, Gl. Åstedvej - Iltindhold i vand",
-      parameter: "Opløst ilt",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 674,
-    },
-    {
-      stationid: 1013,
-      stationname: "Hinnerup Å, Gl. Åstedvej - Temperatur",
-      parameter: "Temperatur",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 674,
-    },
-    {
-      stationid: 726,
-      stationname: "Skive Å, Bilstrupparken - Vandets iltindhold",
-      parameter: "Opløst ilt",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 833,
-    },
-    {
-      stationid: 725,
-      stationname: "Skive Å, Bilstrupparken - Vandtemperatur",
-      parameter: "Temperatur",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 833,
-    },
-    {
-      stationid: 724,
-      stationname: "Skive Å, Lundbro - Vandets iltindhold",
-      parameter: "Opløst ilt",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 678,
-    },
-    {
-      stationid: 723,
-      stationname: "Skive Å, Lundbro - Vandtemperatur",
-      parameter: "Temperatur",
-      loc_owner: "Skive Kommune",
-      alarm: "OK",
-      locid: 678,
-    },
-    {
-      stationid: 757,
-      stationname: "88.1344_3",
-      parameter: "Vandstand",
-      loc_owner: "Gl. Hørning",
-      alarm: "OK",
-      locid: 838,
-    },
-    {
-      stationid: 758,
-      stationname: "Gl. Hørning Natur1",
-      parameter: "Vandstand",
-      loc_owner: "Gl. Hørning",
-      alarm: "OK",
-      locid: 839,
-    },
-    {
-      stationid: 759,
-      stationname: "Gl. Hørning Natur2",
-      parameter: "Vandstand",
-      loc_owner: "Gl. Hørning",
-      alarm: "OK",
-      locid: 840,
-    },
-    {
-      stationid: 756,
-      stationname: "Venø_1",
-      parameter: "Vandstand",
-      loc_owner: "Struer Kommune",
-      alarm: "OK",
-      locid: 837,
-    },
-    {
-      stationid: 760,
-      stationname: "Karstoft B1",
-      parameter: "Vandstand",
-      loc_owner: "Rambøll",
-      alarm: "OK",
-      locid: 841,
-    },
-    {
-      stationid: 1021,
-      stationname: "Rødding Å, Lysgårdvej",
-      parameter: "Vandstand",
-      loc_owner: "Herning Kommune",
-      alarm: "OK",
-      locid: 1294,
-    },
-    {
-      stationid: 1022,
-      stationname: "Tranholm Bæk, Østergade",
-      parameter: "Vandstand",
-      loc_owner: "Herning Kommune",
-      alarm: "OK",
-      locid: 1295,
-    },
-    {
-      stationid: 792,
-      stationname: "Hornslet, Sortevej",
-      parameter: "Vandstand",
-      loc_owner: "Syddjurs Spildevand",
-      alarm: "OK",
-      locid: 848,
-    },
-    {
-      stationid: 188,
-      stationname: "Ryom Å vandstand",
-      parameter: "Vandstand",
-      loc_owner: "Syddjurs Forsyning",
-      alarm: "OK",
-      locid: 366,
-    },
-    {
-      stationid: 1138,
-      stationname: "Logger 370",
-      parameter: "Vandstand",
-      loc_owner: "PC",
-      alarm: "OK",
-      locid: 1343,
-    },
-    {
-      stationid: 1140,
-      stationname: "Logger 374",
-      parameter: "Vandstand",
-      loc_owner: "PC",
-      alarm: "OK",
-      locid: 1343,
-    },
-    {
-      stationid: 1134,
-      stationname: "Logger 439",
-      parameter: "Vandstand",
-      loc_owner: "PC",
-      alarm: "OK",
-      locid: 1343,
-    },
-    {
-      stationid: 1132,
-      stationname: "Logger 483",
-      parameter: "Vandstand",
-      loc_owner: "PC",
-      alarm: "OK",
-      locid: 1343,
-    },
-    {
-      stationid: 1139,
-      stationname: "Logger 486",
-      parameter: "Vandstand",
-      loc_owner: "PC",
-      alarm: "OK",
-      locid: 1343,
-    },
-    {
-      stationid: 1131,
-      stationname: "Logger 496",
-      parameter: "Vandstand",
-      loc_owner: "PC",
-      alarm: "OK",
-      locid: 1343,
-    },
-    {
-      stationid: 1137,
-      stationname: "Logger 497",
-      parameter: "Vandstand",
-      loc_owner: "PC",
-      alarm: "OK",
-      locid: 1343,
-    },
-    {
-      stationid: 1135,
-      stationname: "Logger 499",
-      parameter: "Vandstand",
-      loc_owner: "PC",
-      alarm: "OK",
-      locid: 1343,
-    },
-    {
-      stationid: 1133,
-      stationname: "Logger 500",
-      parameter: "Vandstand",
-      loc_owner: "PC",
-      alarm: "OK",
-      locid: 1343,
-    },
-    {
-      stationid: 796,
-      stationname: "PC2 (496)",
-      parameter: "Vandstand",
-      loc_owner: "PC Brøndboring",
-      alarm: "OK",
-      locid: 849,
-    },
-    {
-      stationid: 797,
-      stationname: "PC3 - (497)",
-      parameter: "Vandstand",
-      loc_owner: "PC Brøndboring",
-      alarm: "OK",
-      locid: 849,
-    },
-    {
-      stationid: 1076,
-      stationname: "PC (370)",
-      parameter: "Vandstand",
-      loc_owner: "PC Brøndboring",
-      alarm: "OK",
-      locid: 849,
-    },
-    {
-      stationid: 1087,
-      stationname: "PC (374)",
-      parameter: "Vandstand",
-      loc_owner: "PC Brøndboring",
-      alarm: "OK",
-      locid: 849,
-    },
-    {
-      stationid: 1077,
-      stationname: "PC (439)",
-      parameter: "Vandstand",
-      loc_owner: "PC Brøndboring",
-      alarm: "OK",
-      locid: 849,
-    },
-    {
-      stationid: 1078,
-      stationname: "PC (483)",
-      parameter: "Vandstand",
-      loc_owner: "PC Brøndboring",
-      alarm: "OK",
-      locid: 849,
-    },
-    {
-      stationid: 1079,
-      stationname: "PC (486)",
-      parameter: "Vandstand",
-      loc_owner: "PC Brøndboring",
-      alarm: "OK",
-      locid: 849,
-    },
-    {
-      stationid: 799,
-      stationname: "PC5 (500)",
-      parameter: "Vandstand",
-      loc_owner: "PC Brøndboring",
-      alarm: "OK",
-      locid: 849,
-    },
-    {
-      stationid: 1098,
-      stationname: "Vandmåler Djurs",
-      parameter: "Flow",
-      loc_owner: "PC",
-      alarm: "OK",
-      locid: 1338,
-    },
-    {
-      stationid: 1026,
-      stationname: "V.Nebel R1(3)",
-      parameter: "Vandstand",
-      loc_owner: "EnviDan",
-      alarm: "OK",
-      locid: 1299,
-    },
-    {
-      stationid: 1027,
-      stationname: "V.Nebel R2(11)",
-      parameter: "Vandstand",
-      loc_owner: "EnviDan",
-      alarm: "OK",
-      locid: 1300,
-    },
-    {
-      stationid: 1028,
-      stationname: "V.Nebel R3(16)",
-      parameter: "Vandstand",
-      loc_owner: "EnviDan",
-      alarm: "OK",
-      locid: 1301,
-    },
-    {
-      stationid: 1029,
-      stationname: "V.Nebel R4(21)",
-      parameter: "Vandstand",
-      loc_owner: "EnviDan",
-      alarm: "OK",
-      locid: 1302,
-    },
-    {
-      stationid: 1030,
-      stationname: "V.Nebel V1",
-      parameter: "Vandstand",
-      loc_owner: "EnviDan",
-      alarm: "OK",
-      locid: 1303,
-    },
-    {
-      stationid: 1031,
-      stationname: "V.Nebel V2",
-      parameter: "Vandstand",
-      loc_owner: "EnviDan",
-      alarm: "OK",
-      locid: 1304,
-    },
-    {
-      stationid: 1093,
-      stationname: "Tinglev_168.1378_1",
-      parameter: "Vandstand",
-      loc_owner: "GEUS",
-      alarm: "OK",
-      locid: 1337,
-    },
-    {
-      stationid: 1094,
-      stationname: "Tinglev_168.1378_2",
-      parameter: "Vandstand",
-      loc_owner: "GEUS",
-      alarm: "OK",
-      locid: 1337,
-    },
-    {
-      stationid: 1095,
-      stationname: "Tinglev_168.1378_3",
-      parameter: "Vandstand",
-      loc_owner: "GEUS",
-      alarm: "OK",
-      locid: 1337,
-    },
-    {
-      stationid: 1096,
-      stationname: "Tinglev_168.1378_4",
-      parameter: "Vandstand",
-      loc_owner: "GEUS",
-      alarm: "OK",
-      locid: 1337,
-    },
-    {
-      stationid: 1097,
-      stationname: "Tinglev_168.1378_5",
-      parameter: "Vandstand",
-      loc_owner: "GEUS",
-      alarm: "OK",
-      locid: 1337,
-    },
-    {
-      stationid: 555,
-      stationname: "Grusvaerk Anders Olsen 21",
-      parameter: null,
-      loc_owner: "Grusværk Anders Olsen",
-      alarm: "OK",
-      locid: 731,
-    },
-    {
-      stationid: 554,
-      stationname: "Grusvaerk Anders Olsen 8",
-      parameter: null,
-      loc_owner: "Grusværk Anders Olsen",
-      alarm: "OK",
-      locid: 730,
-    },
-    {
-      stationid: 553,
-      stationname: "Grusvaerk Anders Olsen 9",
-      parameter: null,
-      loc_owner: "Grusværk Anders Olsen",
-      alarm: "OK",
-      locid: 729,
-    },
-  ],
-  _execution_time: 0.017,
-};
