+++ src/pages/station/ActionArea.js
@@ -1,10 +1,6 @@
 import React, { useState } from "react";
-import { makeStyles, withStyles, useTheme } from "@material-ui/core/styles";
+import { makeStyles, useTheme } from "@material-ui/core/styles";
 import useMediaQuery from "@material-ui/core/useMediaQuery";
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
-import Button from "@material-ui/core/Button";
-import clsx from "clsx";
 import { AddCircle, EditRounded, PhotoCameraRounded } from "@material-ui/icons";
 import BottomNavigation from "@material-ui/core/BottomNavigation";
 import BottomNavigationAction from "@material-ui/core/BottomNavigationAction";
@@ -46,62 +42,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const IconButtonWithText = withStyles({
-  color: "#ffa137",
-  label: {
-    flexDirection: "column",
-  },
-})(Button);
-
-const ColorButton = withStyles((theme) => ({
-  root: {
-    color: theme.palette.getContrastText("#ffa137"),
-    backgroundColor: "#ffa137",
-    "&:hover": {
-      backgroundColor: "#ffa137",
-    },
-  },
-}))(Button);
-
-export default function ActionArea({
-  open,
-  formToShow,
-  setFormToShow,
-  canEdit,
-}) {
-  const classes = useStyles();
-  const [openDialog, setOpenDialog] = useState(false);
-
-  // return (
-  //   <>
-  //     <AppBar
-  //       position='fixed'
-  //       color='inherit'
-  //       className={clsx(classes.appBar1, {
-  //         [classes.appBarShift]: open,
-  //       })}
-  //     >
-  //       <Toolbar style={{ display: "flex", justifyContent: "space-between" }}>
-  //         <ColorButton
-  //           onClick={(e) => {
-  //             setShowForm(true);
-  //           }}
-  //         >
-  //           <AddCircle />
-  //           Indberet pejling
-  //         </ColorButton>
-  //         <ColorButton>
-  //           <EditRounded />
-  //           Ret stamdata
-  //         </ColorButton>
-  //         <ColorButton>
-  //           <PhotoCameraRounded />
-  //           Tag billede
-  //         </ColorButton>
-  //       </Toolbar>
-  //     </AppBar>
-  //   </>
-  // );
+export default function ActionArea({ formToShow, setFormToShow, canEdit }) {
   return (
     <BottomNav
       formToShow={formToShow}
@@ -111,8 +52,8 @@ export default function ActionArea({
   );
 }
 
-function BottomNav({ open, formToShow, setFormToShow, canEdit }) {
-  const [value, setValue] = useState(-1);
+function BottomNav({ setFormToShow, canEdit }) {
+  const [value, _] = useState(-1);
   const classes = useStyles();
   const theme = useTheme();
   const matches = useMediaQuery(theme.breakpoints.down("sm"));
@@ -121,7 +62,6 @@ function BottomNav({ open, formToShow, setFormToShow, canEdit }) {
       className={classes.root}
       value={value}
       onChange={(event, newValue) => {
-        //setValue(newValue);
         if (newValue === 0) {
           setFormToShow("ADDPEJLING");
           setTimeout(() => {
