+++ src/pages/station/CaptureBearing.js
@@ -1,40 +1,7 @@
 import React, { useState } from "react";
-import Dialog from "@material-ui/core/Dialog";
-import Button from "@material-ui/core/Button";
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
-import { makeStyles } from "@material-ui/core/styles";
-import IconButton from "@material-ui/core/IconButton";
-import Typography from "@material-ui/core/Typography";
-import CloseIcon from "@material-ui/icons/Close";
-import Slide from "@material-ui/core/Slide";
 import QrReader from "react-qr-scanner";
 
-const useStyles = makeStyles((theme) => ({
-  appBar: {
-    position: "relative",
-  },
-  title: {
-    marginLeft: theme.spacing(2),
-    flex: 1,
-  },
-}));
-
-const Transition = React.forwardRef(function Transition(props, ref) {
-  return <Slide direction='up' ref={ref} {...props} />;
-});
-
 export default function CaptureBearing() {
-  const classes = useStyles();
-  const [open, setOpen] = React.useState(false);
-
-  const handleClickOpen = () => {
-    setOpen(true);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
   const [result, setResult] = useState("no result");
 
   const handleScan = (data) => {
@@ -45,46 +12,16 @@ export default function CaptureBearing() {
     }
   };
 
-  const camStyle = {
-    // display: "flex",
-    // justifyContent: "center",
-    // marginTop: "-50px",
-  };
+  const camStyle = {};
 
   const previewStyle = {
     height: 300,
     width: 300,
-    //display: "flex",
-    //justifyContent: "center",
   };
 
   const handleError = (error) => console.error(error);
 
   return (
-    // <Dialog
-    //   fullScreen
-    //   open={open}
-    //   onClose={handleClose}
-    //   TransitionComponent={Transition}
-    // >
-    //   <AppBar className={classes.appBar}>
-    //     <Toolbar>
-    //       <IconButton
-    //         edge='start'
-    //         color='inherit'
-    //         onClick={handleClose}
-    //         aria-label='close'
-    //       >
-    //         <CloseIcon />
-    //       </IconButton>
-    //       <Typography variant='h6' className={classes.title}>
-    //         Sound
-    //       </Typography>
-    //       <Button autoFocus color='inherit' onClick={handleClose}>
-    //         save
-    //       </Button>
-    //     </Toolbar>
-    //   </AppBar>
     <div style={camStyle}>
       <QrReader
         delay={100}
@@ -94,6 +31,5 @@ export default function CaptureBearing() {
       />
       <p>{result}</p>
     </div>
-    // </Dialog>
   );
 }
