+++ src/pages/station/CaptureDialog.js
@@ -1,11 +1,9 @@
 import React, { useState } from "react";
 import Dialog from "@material-ui/core/Dialog";
-import Button from "@material-ui/core/Button";
 import AppBar from "@material-ui/core/AppBar";
 import Toolbar from "@material-ui/core/Toolbar";
 import { makeStyles } from "@material-ui/core/styles";
 import IconButton from "@material-ui/core/IconButton";
-import Typography from "@material-ui/core/Typography";
 import CloseIcon from "@material-ui/icons/Close";
 import Slide from "@material-ui/core/Slide";
 import QrReader from "react-qr-scanner";
@@ -27,15 +25,7 @@ const Transition = React.forwardRef(function Transition(props, ref) {
 
 export default function CaptureDialog({ handleClose, open }) {
   const classes = useStyles();
-  //   const [open, setOpen] = React.useState(false);
 
-  //   const handleClickOpen = () => {
-  //     setOpen(true);
-  //   };
-
-  //   const handleClose = () => {
-  //     setOpen(false);
-  //   };
   const [result, setResult] = useState("no result");
 
   const handleScan = (data) => {
@@ -50,7 +40,6 @@ export default function CaptureDialog({ handleClose, open }) {
     display: "flex",
     justifyContent: "center",
     alignItems: "center",
-    // marginTop: "-50px",
     flexDirection: "column",
   };
 
@@ -80,12 +69,6 @@ export default function CaptureDialog({ handleClose, open }) {
           >
             <CloseIcon />
           </IconButton>
-          {/* <Typography variant='h6' className={classes.title}>
-            Sound
-          </Typography> */}
-          {/* <Button autoFocus color='inherit' onClick={handleClose}>
-            save
-          </Button> */}
         </Toolbar>
       </AppBar>
       <div style={camStyle}>
