+++ src/pages/station/DeleteAlert.js
@@ -2,8 +2,6 @@ import React from "react";
 import Button from "@material-ui/core/Button";
 import Dialog from "@material-ui/core/Dialog";
 import DialogActions from "@material-ui/core/DialogActions";
-import DialogContent from "@material-ui/core/DialogContent";
-import DialogContentText from "@material-ui/core/DialogContentText";
 import DialogTitle from "@material-ui/core/DialogTitle";
 
 export default function DeleteAlert({
@@ -13,11 +11,6 @@ export default function DeleteAlert({
   setDialogOpen,
 }) {
   const [open, setOpen] = React.useState(false);
-
-  const handleClickOpen = () => {
-    setOpen(true);
-    setDialogOpen(true);
-  };
 
   const handleClose = () => {
     setOpen(false);
